Move proxy and skeleton containers from methods to common test resources
