+++ score/mw/com/test/common_test_resources/skeleton_container.h
